requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
streamlit
requests
python-dotenv
ollama
